src/main/java/com/samsung/nguyenquanglinh/services/CustomerService.java
package com.samsung.nguyenquanglinh.services;

import com.samsung.nguyenquanglinh.repositories.Customer;
import com.samsung.nguyenquanglinh.repositories.CustomerRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.time.LocalDateTime;
import java.util.List;

@Service
public class CustomerService {

    @Autowired
    private CustomerRepository customerRepository;

    public Customer addCustomer(Customer customer) {
        customer.setCreatedAt(LocalDateTime.now());
        return customerRepository.save(customer);
    }

    public List<Customer> getAllCustomers() {
        return customerRepository.findAll();
    }
}
